Validation updates, ResultSetController, Form Work, sub-forms, picklist updates (Mostly checking in because it is a weekend and I want to make sure my work gets saved)
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@154 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/AttrsDef.java
@@ -15,7 +15,7 @@ public class AttrsDef  implements java.io.Serializable {
     // Fields    
 
      protected Integer attrsDefId;
-     protected Integer disciplineType;
+     protected Short disciplineType;
      protected Short tableType;
      protected Short subType;
      protected String fieldName;
@@ -56,14 +56,14 @@ public class AttrsDef  implements java.io.Serializable {
     /**
      *      *            @hibernate.property
      *             column="disciplineType"
-     *             length="10"
+     *             length="5"
      *         
      */
-    public Integer getDisciplineType() {
+    public Short getDisciplineType() {
         return this.disciplineType;
     }
     
-    public void setDisciplineType(Integer disciplineType) {
+    public void setDisciplineType(Short disciplineType) {
         this.disciplineType = disciplineType;
     }
